+++ docsrc/source/changelog.rst
@@ -9,7 +9,8 @@ v1.2.4
 
 - Add `nltk-rake` support for keyphrases
 - Add more documentation around cluster reporting
-- Enable Dataset and Series access pandas DataFrame and Series methods, respectively
+- Enable `Dataset` and `Series` access `pandas` `DataFrame` and `Series` methods, respectively
+- Change `Dataset.health` from a property to a method and add `pandas` `DataFrame` output
 
 v1.2.3
 -----------